+++ packages/expo-cli/src/commands/credentials.ts
@@ -14,10 +14,11 @@ type Options = {
   };
 };
 
-export default function (program: CommanderStatic) {
+export default function(program: CommanderStatic) {
   program
     .command('credentials:manager')
     .description('Manage your credentials')
+    .helpGroup('credentials')
     .option('-p --platform <platform>', 'Platform: [android|ios]', /^(android|ios)$/i)
     .asyncAction(async (options: Options) => {
       const projectDir = process.cwd();