+++ packages/expo-cli/src/commands/login.ts
@@ -2,11 +2,12 @@ import { Command } from 'commander';
 
 import { login } from '../accounts';
 
-export default function (program: Command) {
+export default function(program: Command) {
   program
     .command('login')
+    .description('Login to an Expo account')
     .alias('signin')
-    .description('Login with your Expo account')
+    .helpGroup('auth')
     .option('-u, --username [string]', 'Username')
     .option('-p, --password [string]', 'Password')
     .asyncAction(login);