+++ packages/expo-cli/src/commands/logout.ts
@@ -12,6 +12,10 @@ async function action() {
   }
 }
 
-export default function (program: Command) {
-  program.command('logout').description('Logout from your Expo account').asyncAction(action);
+export default function(program: Command) {
+  program
+    .command('logout')
+    .description('Logout of an Expo account')
+    .helpGroup('auth')
+    .asyncAction(action);
 }