+++ packages/expo-cli/src/commands/push-creds.ts
@@ -5,8 +5,9 @@ import log from '../log';
 
 export default function (program: Command) {
   program
-    .command('push:android:upload [project-dir]')
-    .description('Uploads a Firebase Cloud Messaging key for Android push notifications.')
+    .command('push:android:upload [path]')
+    .description('Upload an FCM key for Android push notifications')
+    .helpGroup('notifications')
     .option('--api-key [api-key]', 'Server API key for FCM.')
     .asyncActionProjectDir(async (projectDir: string, options: { apiKey?: string }) => {
       if (!options.apiKey || options.apiKey.length === 0) {
@@ -22,8 +23,9 @@ export default function (program: Command) {
     });
 
   program
-    .command('push:android:show [project-dir]')
-    .description('Print the value currently in use for FCM notifications for this project.')
+    .command('push:android:show [path]')
+    .description('Log the value currently in use for FCM notifications for this project')
+    .helpGroup('notifications')
     .asyncActionProjectDir(async (projectDir: string) => {
       const ctx = new Context();
       await ctx.init(projectDir);
@@ -39,8 +41,9 @@ export default function (program: Command) {
     });
 
   program
-    .command('push:android:clear [project-dir]')
-    .description('Deletes a previously uploaded FCM credential.')
+    .command('push:android:clear [path]')
+    .description('Delete a previously uploaded FCM credential')
+    .helpGroup('notifications')
     .asyncActionProjectDir(async (projectDir: string) => {
       const ctx = new Context();
       await ctx.init(projectDir);