+++ packages/expo-cli/src/commands/register.ts
@@ -2,9 +2,10 @@ import { Command } from 'commander';
 
 import { register } from '../accounts';
 
-export default function (program: Command) {
+export default function(program: Command) {
   program
     .command('register')
+    .helpGroup('auth')
     .description('Sign up for a new Expo account')
     .asyncAction(() => register());
 }